refactored mcp to be more modular and best practice
