+++ gpt_researcher/retrievers/__init__.py
@@ -11,7 +11,7 @@ from .serpapi.serpapi import SerpApiSearch
 from .serper.serper import SerperSearch
 from .tavily.tavily_search import TavilySearch
 from .exa.exa import ExaSearch
-from .mcp.mcp_retriever import MCPRetriever
+from .mcp import MCPRetriever
 
 __all__ = [
     "TavilySearch",
