--- gpt_researcher/tools/__init__.py
@@ -1,3 +1,0 @@
-from .mcp_tools import MCPToolsProvider
-
-__all__ = ["MCPToolsProvider"]
